Improve chart loading logic
