+++ src/hooks/useIsomorphicLayoutEffect.tsx
@@ -1,0 +1,6 @@
+import { useEffect, useLayoutEffect } from 'react';
+
+const useIsomorphicLayoutEffect =
+  typeof window !== 'undefined' ? useLayoutEffect : useEffect;
+
+export default useIsomorphicLayoutEffect;
