+++ src/pages/ChainStats/components/ChartCoin/index.ts
@@ -1,0 +1,2 @@
+export * from './ChartCoin';
+export { default } from './ChartCoin';
