+++ src/pages/ChainStats/components/ChartDurationPicker/index.ts
@@ -1,0 +1,2 @@
+export * from './ChartDurationPicker';
+export { default } from './ChartDurationPicker';
