+++ src/pages/ChainStats/components/ChartMetrics/index.ts
@@ -1,0 +1,2 @@
+export * from './ChartMetrics';
+export { default } from './ChartMetrics';
