+++ src/pages/ChainStats/components/index.ts
@@ -1,0 +1,3 @@
+export * from './ChartCoin';
+export * from './ChartDurationPicker';
+export * from './ChartMetrics';
